// src/main/frontend/src/enums.ts
export enum NodeType {
    Provenance = 'provenance',
    Resource = 'resource',
}

export enum DetailType {
    Who = 'who',
    Where = 'where',
    When = 'when',
    HowSoftware = 'howSoftware',
    HowInit = 'howInit',
    HowDelta = 'howDelta',
    WhyMotivation = 'whyMotivation',
    WhyProvenanceSchema = 'whyProvenanceSchema',
}